# docs/swagger.yaml
definitions:
...
  db.Interest:
    properties:
      name:
        type: string
    type: object
  golaerror.Error:
    properties:
      additionalData:
        $ref: '#/definitions/golaerror.additionalData'
      errorCode:
        type: string
      errorMessage:
        type: string
    type: object
  golaerror.additionalData:
    type: object
...
  models.JSONString:
    properties:
      types.JSONText:
        items:
          type: integer
        type: array
    type: object
...
  request.PostURIRequest:
    properties:
      postUID:
        type: string
    required:
    - postUID
    type: object
...
  request.PublishRequest:
    properties:
      draft_id:
        type: string
    required:
    - draft_id
    type: object
  request.SearchInterests:
    properties:
      searchKeyword:
        type: string
      selectedTags:
        items:
          type: string
        type: array
    required:
    - selectedTags
    type: object
...
  response.LikedByCount:
    properties:
      like_count:
        type: integer
    type: object
  response.Post:
    properties:
      author_id:
        type: string
      author_name:
        type: string
      comment_count:
        type: integer
      interests:
        items:
          type: string
        type: array
      is_viewer_followed_author:
        type: boolean
      is_viewer_is_author:
        type: boolean
      is_viewer_liked:
        type: boolean
      like_count:
        type: integer
      post_data:
        $ref: '#/definitions/models.JSONString'
      post_id:
        type: string
      preview_image:
        type: string
      published_at:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/post/v1/post/:post_id:
    get:
      consumes:
      - application/json
      description: get a post
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PostURIRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: GetPost
      tags:
      - post
  /api/post/v1/post/:post_id/like:
    get:
      consumes:
      - application/json
      description: like a post
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PostURIRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LikedByCount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: Like
      tags:
      - post
  /api/post/v1/post/:post_id/unlike:
    get:
      consumes:
      - application/json
      description: unlike a post
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PostURIRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LikedByCount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: Unlike
      tags:
      - post
  /api/post/v1/post/comment:
    post:
      consumes:
      - application/json
      description: get all interests or search a interest with keyword
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SearchInterests'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Interest'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: GetInterests
      tags:
      - interest
  /api/post/v1/post/publish:
    post:
      consumes:
      - application/json
      description: publish a draft to post
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PublishRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/golaerror.Error'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: PublishPost
      tags:
      - post
swagger: "2.0"